07.Chatbot_RAG_customKB_LangChain/requirements.txt
chromadb==1.5.0
langchain==0.0.338
python-dateutil==2.8.2
streamlit-chat==0.1.1
colorama==0.4.6
openai==2.17.0
python-dotenv==1.2.1
tiktoken==0.12.0
streamlit==1.54.0
requests==2.31.0
Pillow==12.1.0
whisper==1.1.10